+++ pyron-engine/src/main/scala/com/cloudentity/pyron/apigroup/ApiGroup.scala
@@ -4,5 +4,10 @@ import com.cloudentity.pyron.domain.flow.GroupMatchCriteria
 import com.cloudentity.pyron.domain.rule.RuleConfWithPlugins
 import com.cloudentity.pyron.rule.Rule
 
-case class ApiGroup(matchCriteria: GroupMatchCriteria, rules: List[Rule])
-case class ApiGroupConf(matchCriteria: GroupMatchCriteria, rules: List[RuleConfWithPlugins])
+case class ApiGroupId(value: String)
+case class ApiGroup(id: ApiGroupId, matchCriteria: GroupMatchCriteria, rules: List[Rule])
+case class ApiGroupConf(id: ApiGroupId, matchCriteria: GroupMatchCriteria, rules: List[RuleConfWithPlugins])
+
+object ApiGroup {
+  val propertiesKey = "_internal.apiGroup"
+}