+++ pyron-engine/src/main/scala/com/cloudentity/pyron/apigroup/ApiGroupsStore.scala
@@ -130,15 +130,19 @@ class ApiGroupsStoreVerticle extends ScalaServiceVerticle with ApiGroupsStore {
       case (Some(rules), None) =>
         // fallback for deprecated Rules configuration
         val rulesValue = Try(new JsonArray(rules.noSpaces)).orElse(Try(new JsonObject(rules.noSpaces))).toOption.getOrElse(null)
-        buildApiGroups(new JsonObject().put("_rules", rulesValue), defaultProxyRulesOpt, true)
+
+        val defaultGroupMatchingCriteria = new JsonObject().put("basePath", "")
+        val defaultApiGroup = new JsonObject().put("_group", defaultGroupMatchingCriteria).put("_rules", rulesValue)
+
+        buildApiGroups(new JsonObject().put("default", defaultApiGroup), defaultProxyRulesOpt)
       case (None, Some(apiGroups)) =>
-        buildApiGroups(apiGroups, defaultProxyRulesOpt, false)
+        buildApiGroups(apiGroups, defaultProxyRulesOpt)
       case (Some(_), Some(_)) =>
         Future.failed(new Exception("Both API Groups and Rules can't be defined. Move Rules to Api Groups."))
     }
   }
 
-  private def buildApiGroups(apiGroupsConf: JsonObject, defaultProxyRulesOpt: Option[Json], deprecatedRules: Boolean): Future[(List[ApiGroup], List[ApiGroupConf])] =
+  private def buildApiGroups(apiGroupsConf: JsonObject, defaultProxyRulesOpt: Option[Json]): Future[(List[ApiGroup], List[ApiGroupConf])] =
     ApiGroupReader.readApiGroupLevels(apiGroupsConf.toString) match {
       case ValidResult(_, groupLevel) =>
         val unresolvedGroupResults: List[ReadResult[ApiGroupConfUnresolved]] =
@@ -162,7 +166,8 @@ class ApiGroupsStoreVerticle extends ScalaServiceVerticle with ApiGroupsStore {
 
             rulesStore.decodeRules(loggingTag(group), group.value.rules, defaultProxyRulesOpt, pluginIds).toScala()
               .map { case (rs, rsConfs) =>
-                ValidResult(group.path, (ApiGroup(group.value.matchCriteria, rs), ApiGroupConf(group.value.matchCriteria, rsConfs)))
+                val id = ApiGroupId(group.path.reverse.mkString("."))
+                ValidResult(group.path, (ApiGroup(id, group.value.matchCriteria, rs), ApiGroupConf(id, group.value.matchCriteria, rsConfs)))
               }
               .recover { case ex: Throwable =>
                 InvalidResult[(ApiGroup, ApiGroupConf)](group.path, ex.getMessage)